Moved LogActionRegistry into wagtailcore
Since logs are generated and stored in wagtailcore, I think it makes
sense for the list of available actions to be managed there too.
