+++ wagtail/core/tests/test_audit_log.py
@@ -6,6 +6,7 @@ from django.test import TestCase
 from django.utils import timezone
 from freezegun import freeze_time
 
+from wagtail.core.logging import LogActionRegistry
 from wagtail.core.models import (
     Page, PageLogEntry, PageViewRestriction, Task, Workflow, WorkflowTask)
 from wagtail.tests.testapp.models import SimplePage
@@ -288,3 +289,32 @@ class TestAuditLog(TestCase):
         restriction.restriction_type = PageViewRestriction.PASSWORD
         restriction.save()
         self.assertEqual(PageLogEntry.objects.filter(action='wagtail.view_restriction.edit').count(), 1)
+
+
+def test_hook(actions):
+    return actions.register_action('test.custom_action', 'Custom action', 'Tested!')
+
+
+class TestAuditLogHooks(TestCase, WagtailTestUtils):
+    def setUp(self):
+        self.root_page = Page.objects.get(id=2)
+
+    def test_register_log_actions_hook(self):
+        # testapp/wagtail_hooks.py defines a 'blockquote' rich text feature with a hallo.js
+        # plugin, via the register_rich_text_features hook; test that we can retrieve it here
+        log_actions = LogActionRegistry()
+        actions = log_actions.get_actions()
+        self.assertIn('wagtail.create', actions)
+
+    def test_action_format_message(self):
+        log_entry = PageLogEntry.objects.log_action(self.root_page, action='test.custom_action')
+
+        log_actions = LogActionRegistry()
+        self.assertEqual(log_actions.format_message(log_entry), "Unknown test.custom_action")
+        self.assertNotIn('test.custom_action', log_actions.get_actions())
+
+        with self.register_hook('register_log_actions', test_hook):
+            log_actions = LogActionRegistry()
+            self.assertIn('test.custom_action', log_actions.get_actions())
+            self.assertEqual(log_actions.format_message(log_entry), "Tested!")
+            self.assertEqual(log_actions.get_action_label('test.custom_action'), 'Custom action')
